Update statistics precision

+++ src/Data/NASRAC.Models.Game/Stats/Abstractions/BaseSessionStats.cs
@@ -20,12 +20,12 @@ public abstract class BaseSessionStats : BaseStats, IBaseSessionStats
     /// <summary>
     /// Driver's overall average position
     /// </summary>
-    public int AveragePosition { get; set; }
+    public double AveragePosition { get; set; }
     
     /// <summary>
     /// Driver's average position while on the lead lap
     /// </summary>
-    public int AverageRunningPosition { get; set; }
+    public double AverageRunningPosition { get; set; }
     
     /// <summary>
     /// How many laps the driver was in the top 15 standings
@@ -35,7 +35,7 @@ public abstract class BaseSessionStats : BaseStats, IBaseSessionStats
     /// <summary>
     /// Percentage of laps the driver was in the top 15 standings
     /// </summary>
-    public int Top15LapPercentage { get; set; }
+    public double Top15LapPercentage { get; set; }
     
     /// <summary>
     /// How many laps the driver led
@@ -45,7 +45,7 @@ public abstract class BaseSessionStats : BaseStats, IBaseSessionStats
     /// <summary>
     /// Percentage of laps the driver led
     /// </summary>
-    public int LapLedPercentage { get; set; }
+    public double LapLedPercentage { get; set; }
 
     /// <summary>
     /// How many laps the driver spent under caution
@@ -55,7 +55,7 @@ public abstract class BaseSessionStats : BaseStats, IBaseSessionStats
     /// <summary>
     /// Percentage of laps the driver spent under caution
     /// </summary>
-    public int CautionLapPercentage { get; set; }
+    public double CautionLapPercentage { get; set; }
 
     /// <summary>
     /// How many cautions the driver caused
